atualizei exercicios

+++ exercicios/ex004.py
@@ -1,4 +1,4 @@
-metro = int(input('digite o valor em metros: '))
+metro = float(input('digite o valor em metros: '))
 centimetro = metro * 100
 milimetro = metro * 1000
 print(f'{metro} metros é igual a {centimetro} centimetros e {milimetro} milimetros')
